+++ src/guile/ares/suitbl/presets.scm
@@ -2,7 +2,7 @@
 ;; Copyright © 2026 Andrew Tropin <andrew@trop.in>
 
 (define-module (ares suitbl presets)
-  #:use-module ((ares suitbl definitions) #:select (test-runner*))
+  #:use-module ((ares suitbl definitions) #:select (current-test-runner))
   #:use-module ((ares suitbl runner) #:prefix runner:)
   #:use-module ((ares suitbl state)
                 #:select (set-runner-config-value!
@@ -22,7 +22,7 @@
 ;;; Helpers
 ;;;
 
-(define* (runner->state #:optional (runner (test-runner*)))
+(define* (runner->state #:optional (runner (current-test-runner)))
   (runner:get-state runner))
 
 
@@ -31,32 +31,32 @@
 ;;; Presets
 ;;;
 
-(define* (only-slow! #:optional (runner (test-runner*)))
+(define* (only-slow! #:optional (runner (current-test-runner)))
   "Configure RUNNER to schedule only slow tests."
   (let ((state (runner->state runner)))
     (set-runner-config-value! state 'schedule-tests scheduler:slow)
     (get-runner-config state)))
 
-(define* (only-fast! #:optional (runner (test-runner*)))
+(define* (only-fast! #:optional (runner (current-test-runner)))
   "Configure RUNNER to schedule only fast tests."
   (let ((state (runner->state runner)))
     (set-runner-config-value! state 'schedule-tests scheduler:fast)
     (get-runner-config state)))
 
-(define* (matching! pattern #:optional (runner (test-runner*)))
+(define* (matching! pattern #:optional (runner (current-test-runner)))
   "Configure RUNNER to schedule only tests matching regexp PATTERN."
   (let ((state (runner->state runner)))
     (set-runner-config-value! state 'schedule-tests (scheduler:make-matching pattern))
     (get-runner-config state)))
 
-(define* (rerun-failed-or-all! #:optional (runner (test-runner*)))
+(define* (rerun-failed-or-all! #:optional (runner (current-test-runner)))
   "Configure RUNNER to schedule only tests that failed or errored
 in the previous run."
   (let ((state (runner->state runner)))
     (set-runner-config-value! state 'schedule-tests scheduler:failed-or-all)
     (get-runner-config state)))
 
-(define* (raise-on-error! #:optional (runner (test-runner*)))
+(define* (raise-on-error! #:optional (runner (current-test-runner)))
   "Configure RUNNER to re-raise exceptions on assertion failures and
 errors without unwinding the stack, so the IDE can bring up a stack
 trace viewer."
@@ -64,7 +64,7 @@ trace viewer."
     (set-runner-config-value! state 're-raise? #t)
     (get-runner-config state)))
 
-(define* (reset! #:optional (runner (test-runner*)))
+(define* (reset! #:optional (runner (current-test-runner)))
   "Remove the schedule-tests filter from RUNNER, restoring default
 behavior of running all loaded tests."
   (let ((state (runner->state runner)))