+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The long lines are not hard-wrapped.  Use softwrapping in your editor for readin
 ### Changed
 - Migrated bencode tests from SRFI-64 to suitbl.
 - suitbl runner assertion message type was renamed from `runner/run-assert` to `runner/run-assertion`.
+- suitbl `test-runner*` parameter was renamed to `current-test-runner`.
 - suitbl tests now use `(test DESCRIPTION () BODY ...)` or `(test DESCRIPTION (CONTEXT) BODY ...)`, depending on whether the test context is needed.
 - suitbl test entities now use `test/body-procedure` one-argument procedures instead of `test/body-thunk` thunks.
 - suitbl `test-thunk` was renamed to `test-loader`; `test-thunk` remains as a deprecated alias.